Don't save screenshot if option is disabled
Previously, a screenshot was still taken and saved to file
if option was disabled.

+++ src/sentry/native/native_sdk.cpp
@@ -50,6 +50,10 @@ void sentry_event_set_context(sentry_value_t p_event, const char *p_context_name
 }
 
 inline void _save_screenshot() {
+	if (!SentryOptions::get_singleton()->is_attach_screenshot_enabled()) {
+		return;
+	}
+
 	String screenshot_path = "user://" _SCREENSHOT_FN;
 	DirAccess::remove_absolute(screenshot_path);
